+++ src/js/main.js
@@ -3,5 +3,8 @@ import "./dp/cookies-banner";
 import "./dp/set-display";
 import "./dp/viewport-size";
 import "./dp/nav";
-import "./dp/improve-this-page";
 import "./dp/search";
+import "./dp/improve-this-page";
+import "./dp/improve-this-page-form";
+
+
